Add functionality of using 32-byte passphrase
I was fascinated with ecryptfs and annoyed by the error message given out when $KEYBYTES was set to 32 arbitrarily. Finally I found line 402 in my proposal is the key. Because the number of bytes of "auth tok" is always 16 no matter how many bytes the passphrase is, the "$KEYBYTES,$KEYBYTES" should be replaced with 16.

I have made a few tests and seem working fine. The new 32-byte passphrase has been proved by ecryptfs-unwrap-passphrase.

P.S. This is my first proposal in GitHub ~~ Hope that you can accept that.